# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Whisper Asr Webservice","description":"Whisper ASR Webservice is a general-purpose speech recognition webservice.","version":"1.10.0-dev","contact":{"url":"https://github.com/ahmetoner/whisper-asr-webservice/"},"license":{"name":"MIT License","url":"https://github.com/ahmetoner/whisper-asr-webservice/blob/main/LICENCE"}},"paths":{"/asr":{"post":{"tags":["Endpoints"],"summary":"Asr Task","operationId":"asr_task_asr_post","parameters":[{"name":"encode","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Encode"},"description":"Encode audio first through ffmpeg"},{"name":"task","in":"query","required":false,"schema":{"anyOf":[{"type":"string","enum":["transcribe","translate"]},{"type":"null"}],"title":"Task","default":"transcribe"}},{"name":"language","in":"query","required":false,"schema":{"anyOf":[{"type":"string","enum":["af","am","ar","as","az","ba","be","bg","bn","bo","br","bs","ca","cs","cy","da","de","el","en","es","et","eu","fa","fi","fo","fr","gl","gu","ha","haw","he","hi","hr","ht","hu","hy","id","is","it","ja","jw","ka","kk","km","kn","ko","la","lb","ln","lo","lt","lv","mg","mi","mk","ml","mn","mr","ms","mt","my","ne","nl","nn","no","oc","pa","pl","ps","pt","ro","ru","sa","sd","si","sk","sl","sn","so","sq","sr","su","sv","sw","ta","te","tg","th","tk","tl","tr","tt","ug","uk","ur","uz","vi","yi","yo","zh"]},{"type":"null"}],"title":"Language"}},{"name":"initial_prompt","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Initial Prompt"}},{"name":"vad_filter","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Vad Filter","description":"Enable the voice activity detection (VAD) to filter out parts of the audio without speech"},"include_in_schema":false},{"name":"word_timestamps","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Word Timestamps","description":"Word level timestamps"},"include_in_schema":false},{"name":"diarize","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Diarize","description":"Diarize the input"},"include_in_schema":false},{"name":"min_speakers","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Min Speakers","description":"Min speakers in this file"},"include_in_schema":false},{"name":"max_speakers","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Speakers","description":"Max speakers in this file"},"include_in_schema":false},{"name":"output","in":"query","required":false,"schema":{"anyOf":[{"type":"string","enum":["txt","vtt","srt","tsv","json"]},{"type":"null"}],"title":"Output","default":"txt"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"audio_file":{"type":"string","format":"binary"}},"required":["audio_file"],"title":"Body_asr_task_asr_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/asr/result/{task_id}":{"get":{"tags":["Endpoints"],"summary":"Get Task Result","description":"Get the result of a transcription task.","operationId":"get_task_result_asr_result__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/detect-language":{"post":{"tags":["Endpoints"],"summary":"Detect Language","operationId":"detect_language_detect_language_post","parameters":[{"name":"encode","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Encode"},"description":"Encode audio first through FFmpeg"}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"audio_file":{"type":"string","format":"binary"}},"required":["audio_file"],"title":"Body_detect_language_detect_language_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"type":"object","properties":{"detail":{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"},"title":"Detail"}},"title":"HTTPValidationError"},"ValidationError":{"type":"object","properties":{"loc":{"type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"required":["loc","msg","type"],"title":"ValidationError"}}}}